requirements.txt
cbor2==5.2.0
certifi==2020.12.5
click==7.1.2
Deprecated==1.2.10
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
luma.core==2.2.0
luma.led-matrix==1.5.0
MarkupSafe==1.1.1
Pillow==8.0.1
pkg-resources==0.0.0
pyftdi==0.52.0
pygame==2.0.0
pyserial==3.5
pyusb==1.1.0
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
simpleaudio==1.0.4
smbus2==0.4.0
spidev==3.5
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
ws2812==1.0.0
WTForms==2.3.3
